refactor: 필터 로직을 관리하는 useProductFilter 커스텀 훅 분리
